Split free*.php Laravel dumps into a real directory tree
free3.php / free2.php / free.php は、いずれも複数ファイル分のコードを
1 ファイルに連結したもので、2 つ目以降の `<?php` により Parse error に
なっていた。

  PHP Parse error: syntax error, unexpected token "<" in free.php  on line 77
  PHP Parse error: syntax error, unexpected token "<" in free2.php on line 43
  PHP Parse error: syntax error, unexpected token "<" in free3.php on line 24

中身は同じ Lodge アプリを段階的に育てた 3 世代だったため、
Laravel/lodge-app/step{1,2,3}/ として Laravel の実配置へ分割。

  step1 (旧 free3.php):  6 ファイル  一覧・詳細のみ
  step2 (旧 free2.php): 10 ファイル  リレーションと登録フォーム
  step3 (旧 free.php ): 22 ファイル  FormRequest + Policy まで含む完成形

ファイル中の `// app/Http/Controllers/LodgeController.php` といった
パスコメントをそのまま分割位置として使用しており、コードの中身は変更していない。
分割後は PHP ファイルがすべて php -l を通る。
