+++ Laravel/lodge-app/step1/app/Models/Lodge.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Lodge extends Model
+{
+    protected $fillable = ['name', 'location', 'founded_year'];
+
+    public function members()
+    {
+        return $this->hasMany(Member::class);
+    }
+}
