+++ Laravel/lodge-app/step3/app/Http/Requests/StoreEventRequest.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace App\Http\Requests;
+
+use App\Models\Event;
+use Illuminate\Foundation\Http\FormRequest;
+
+class StoreEventRequest extends FormRequest
+{
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    public function rules(): array
+    {
+        return [
+            'title' => ['required', 'string', 'max:255'],
+            'description' => ['required', 'string'],
+            'date' => ['required', 'date', 'after_or_equal:today'],
+            'type' => ['required', 'string', 'in:' . implode(',', Event::TYPES)],
+        ];
+    }
+}
