+++ Laravel/lodge-app/step3/app/Models/Symbol.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+
+class Symbol extends Model
+{
+    use HasFactory;
+    
+    protected $fillable = ['name', 'description', 'meaning'];
+}
